[FE] refactor: #123 calendarSlicefile 생성 & default slice file에서 calendar 관련 store 정보 삭제
calendar 관련 store 정보를 calendarSlice file로 분리했습니다.

default slice file에서 calendar 관련 store 정보를 삭제했습니다.
